+++ aqueduct/lib/src/http/serializable.dart
@@ -15,6 +15,9 @@ abstract class Serializable {
   /// will be of type 'object'. Each instance variable declared in [serializable] will be a property of this object.
   /// Instance variables are documented according to [APIComponentDocumenter.documentVariable]. See the API reference
   /// for this method for supported types.
+  ///
+  /// You may override the default behavior: f [serializable] implements a static [document] method
+  /// with the same signature of this method, the 'overridden' will be invoked instead of the default behavior.
   static APISchemaObject document(
       APIDocumentContext context, Type serializable) {
     final mirror = reflectClass(serializable);
@@ -23,20 +26,48 @@ abstract class Serializable {
           "Cannot document '${MirrorSystem.getName(mirror.simpleName)}' as 'Serializable', because it is not an 'Serializable'.");
     }
 
+    final overridden = mirror.staticMembers[#document];
+    if (overridden != null) {
+      final isValidInvocation = overridden.returnType.isAssignableTo(reflectType(APISchemaObject)) &&
+        overridden.parameters.length == 2 &&
+        overridden.parameters.first.type
+          .isAssignableTo(reflectType(APIDocumentContext)) &&
+        overridden.parameters.last.type.isAssignableTo(reflectType(Type));
+
+      if (isValidInvocation) {
+        return mirror.invoke(#document, [context, serializable]).reflectee as APISchemaObject;
+      } else {
+        throw StateError("Type '$serializable' provides static method 'document', "
+          "but does not have same signature as 'Serializable.document'.");
+      }
+    }
+
+    String failureReason;
     final properties = <String, APISchemaObject>{};
-    for (final property
-        in mirror.declarations.values.whereType<VariableMirror>()) {
-      properties[MirrorSystem.getName(property.simpleName)] =
-          APIComponentDocumenter.documentVariable(context, property);
+    try {
+      for (final property
+          in mirror.declarations.values.whereType<VariableMirror>()) {
+        properties[MirrorSystem.getName(property.simpleName)] =
+            APIComponentDocumenter.documentVariable(context, property);
+      }
+    } catch (e) {
+      failureReason = e.toString();
     }
 
     final obj = APISchemaObject.object(properties);
-    context.defer(() async {
-      final docs = await DocumentedElement.get(serializable);
-      obj
-        ..title = docs?.summary
-        ..description = docs?.description;
-    });
+    if (failureReason == null) {
+      context.defer(() async {
+        final docs = await DocumentedElement.get(serializable);
+        obj
+          ..title = docs?.summary
+          ..description = docs?.description;
+      });
+    } else {
+      obj.additionalPropertyPolicy = APISchemaAdditionalPropertyPolicy.freeForm;
+      obj.title =
+          "Failed to auto-document type '${MirrorSystem.getName(mirror.simpleName)}'. See description for error.";
+      obj.description = failureReason;
+    }
 
     return obj;
   }